feat: optimise chargen permalink, fix edge stat
Further optimised the chargen permalink data - now everything is a number
in an array - no strings, no keys. Stat editing was broken if the stat
was 0 (which is often the case for starting edge values). Also tweaked
the styles for the v-chips used for abilities so they are now variable
height for longer named abilities in narrow views.