fix: add missing data for destiny types
The destiny types were relying on old code that used to merge duplicate
types together, but that no longer happens and so the types bar entries
just broke. As a quick fix without checking to see if there are any
actual differences, have copied the corebook versions into the destiny
ones.